accelerometer values are floats and in G, not milli-G

+++ fcb-source/fcb-drivers/BSP/Components/Common/accelero.h
@@ -76,7 +76,7 @@ typedef struct
   void      (*ClearIT)(void);
   void      (*FilterConfig)(uint8_t);
   void      (*FilterCmd)(uint8_t);
-  void      (*GetXYZ)(int16_t *);
+  void      (*GetXYZ)(float *);
 }ACCELERO_DrvTypeDef;
